Add FileUploader trait
The FileUploader trait marks an object that can be used to upload a file to a TelegramRequest, that file may be injected in many ways, and that upload is valid in only a certain number of situations. It is expected that the method explicits what those situations are by using concrete types or trait types.